requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=-
#
aiosmtplib==2.0.2
    # via fastapi-mail
alembic==1.12.0
    # via -r requirements.in
annotated-types==0.5.0
    # via pydantic
anyio==3.7.1
    # via
    #   -r requirements.in
    #   fastapi
    #   httpcore
    #   starlette
bcrypt==4.0.1
    # via passlib
blinker==1.6.2
    # via fastapi-mail
certifi==2023.7.22
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.0
    # via requests
click==8.1.7
    # via uvicorn
coverage[toml]==7.3.2
    # via pytest-cov
cryptography==41.0.4
    # via python-jose
dnspython==2.4.2
    # via email-validator
dramatiq[rabbitmq,watch]==1.14.2
    # via -r requirements.in
ecdsa==0.18.0
    # via python-jose
email-validator==2.0.0.post2
    # via fastapi-mail
exceptiongroup==1.1.3
    # via
    #   anyio
    #   pytest
fastapi==0.103.2
    # via -r requirements.in
fastapi-mail==1.4.1
    # via -r requirements.in
gevent==23.9.1
    # via watchdog-gevent
greenlet==3.0.0
    # via
    #   gevent
    #   sqlalchemy
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==0.18.0
    # via httpx
httpx==0.25.0
    # via -r requirements.in
idna==3.4
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
iniconfig==2.0.0
    # via pytest
jinja2==3.1.2
    # via
    #   -r requirements.in
    #   fastapi-mail
mako==1.2.4
    # via alembic
markupsafe==2.1.3
    # via
    #   jinja2
    #   mako
packaging==23.2
    # via pytest
passlib[bcrypt]==1.7.4
    # via -r requirements.in
pika==1.3.2
    # via dramatiq
pluggy==1.3.0
    # via pytest
prometheus-client==0.17.1
    # via dramatiq
psycopg2-binary==2.9.8
    # via -r requirements.in
py-automapper==1.2.3
    # via -r requirements.in
pyasn1==0.5.0
    # via
    #   python-jose
    #   rsa
pycparser==2.21
    # via cffi
pydantic==2.4.2
    # via
    #   -r requirements.in
    #   fastapi
    #   fastapi-mail
    #   pydantic-settings
pydantic-core==2.10.1
    # via pydantic
pydantic-settings==2.0.3
    # via fastapi-mail
pytest==7.4.2
    # via
    #   -r requirements.in
    #   pytest-cov
pytest-cov==4.1.0
    # via -r requirements.in
python-dotenv==1.0.0
    # via
    #   -r requirements.in
    #   pydantic-settings
python-jose[cryptography]==3.3.0
    # via -r requirements.in
requests==2.31.0
    # via -r requirements.in
rsa==4.9
    # via python-jose
schedule==1.2.1
    # via -r requirements.in
six==1.16.0
    # via ecdsa
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
sqlalchemy==2.0.21
    # via
    #   -r requirements.in
    #   alembic
starlette==0.27.0
    # via
    #   fastapi
    #   fastapi-mail
tomli==2.0.1
    # via
    #   coverage
    #   pytest
typeguard==4.1.5
    # via -r requirements.in
typing-extensions==4.8.0
    # via
    #   alembic
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typeguard
    #   uvicorn
urllib3==2.0.6
    # via requests
uvicorn==0.23.2
    # via -r requirements.in
watchdog==3.0.0
    # via
    #   dramatiq
    #   watchdog-gevent
watchdog-gevent==0.1.1
    # via dramatiq
zope-event==5.0
    # via gevent
zope-interface==6.0
    # via gevent
pdfkit==1.0.0
# The following packages are considered to be unsafe in a requirements file:
# setuptools
